Issue #17168: Moved Size Module Inputs to Compilable Folder
