+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/sizes/ExecutableStatementCountCheckTest.java
@@ -184,7 +184,7 @@ public class ExecutableStatementCountCheckTest
         };
 
         verifyWithInlineConfigParser(
-                getNonCompilablePath("InputExecutableStatementCountRecords.java"),
+                getPath("InputExecutableStatementCountRecords.java"),
                 expected);
     }
