+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/sizes/MethodCountCheckTest.java
@@ -199,7 +199,7 @@ public class MethodCountCheckTest extends AbstractModuleTestSupport {
         };
 
         verifyWithInlineConfigParser(
-                getNonCompilablePath("InputMethodCountRecords.java"), expected);
+                getPath("InputMethodCountRecords.java"), expected);
     }
 
 }
