+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/sizes/executablestatementcount/InputExecutableStatementCountRecords.java
@@ -1,19 +1,18 @@
 /*
-MethodLength
-max = 2
-countEmpty = (default)true
-tokens = (default)METHOD_DEF, CTOR_DEF, COMPACT_CTOR_DEF
+ExecutableStatementCount
+max = 1
+tokens = (default)CTOR_DEF, METHOD_DEF, INSTANCE_INIT, STATIC_INIT, COMPACT_CTOR_DEF, LAMBDA
 
 
 */
 
-//non-compiled with javac: Compilable with Java17
-package com.puppycrawl.tools.checkstyle.checks.sizes.methodlength;
+// Java17
+package com.puppycrawl.tools.checkstyle.checks.sizes.executablestatementcount;
 
-public class InputMethodLengthRecordsAndCompactCtors {
+public class InputExecutableStatementCountRecords {
 
     record MyTestRecord() {
-        static {
+        static { // violation
             System.out.println("test");
             System.out.println("test");
             System.out.println("test");
@@ -22,8 +21,7 @@ public class InputMethodLengthRecordsAndCompactCtors {
 
     //compact ctor
     record MyTestRecord2() {
-        // violation below 'Method MyTestRecord2 length is 6 lines (max allowed is 2)'
-        public MyTestRecord2 {
+        public MyTestRecord2 { // violation
             System.out.println("test");
             System.out.println("test");
             System.out.println("test");
@@ -32,7 +30,7 @@ public class InputMethodLengthRecordsAndCompactCtors {
     }
 
     record MyTestRecord3(String str) {
-        void foo() { // violation 'Method foo length is 5 lines (max allowed is 2)'
+        void foo() { // violation
             System.out.println("test");
             System.out.println("test");
             System.out.println("test");
@@ -40,8 +38,7 @@ public class InputMethodLengthRecordsAndCompactCtors {
     }
 
     record MyTestRecord4(int x, int y) {
-        // violation below 'Method MyTestRecord4 length is 7 lines (max allowed is 2)'
-        public MyTestRecord4() {
+        public MyTestRecord4() { // violation
             this(4, 5);
             System.out.println("test");
             System.out.println("test");
@@ -51,7 +48,7 @@ public class InputMethodLengthRecordsAndCompactCtors {
     }
 
     record MyTestRecord5() {
-        static {
+        static { // violation
             int y = 5;
             int z = 10;
             String newString = String.valueOf(y);
@@ -62,10 +59,10 @@ public class InputMethodLengthRecordsAndCompactCtors {
     }
 
     class LocalRecordHelper {
-        Class<?> m(int x) { // violation 'Method m length is 15 lines (max allowed is 2)'
+        Class<?> m(int x) {
             record R76(int x) {
 
-                public R76 { // violation 'Method R76 length is 8 lines (max allowed is 2)'
+                public R76 { // violation
                     int y = 5;
                     int z = 10;
                     String newString = String.valueOf(y);
@@ -78,4 +75,20 @@ public class InputMethodLengthRecordsAndCompactCtors {
             return R76.class;
         }
     }
+
+    private int id(int i) {
+        return i;
+    }
+
+    private final int value = 2;
+
+    private final int field = id(switch(value) {
+        case 0 -> -1;
+        case 2 -> {
+            int temp = 0;
+            temp += 3;
+            yield temp;
+        }
+        default -> throw new IllegalStateException();
+    });
 }
