+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/sizes/linelength/InputLineLengthIgnoringPackageStatements.java
@@ -6,13 +6,8 @@ max = 75
 
 
 */
-//non-compiled with javac: Compilable with Java17
-package com.puppycrawl.tools.   checkstyle.checks.sizes.linelength;
-
-// violation below 'longer than 75 characters (found 76)'
-import java.   security.    interfaces.          RSAMultiPrimePrivateCrtKey;
-
-import java.util.Arrays;
+// Java17
+package com.puppycrawl.tools.checkstyle.checks.sizes.linelength;
 
 public class InputLineLengthIgnoringPackageStatements {
     @Override
