+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/sizes/methodlength/InputMethodLengthRecordsAndCompactCtors.java
@@ -1,18 +1,19 @@
 /*
-ExecutableStatementCount
-max = 1
-tokens = (default)CTOR_DEF, METHOD_DEF, INSTANCE_INIT, STATIC_INIT, COMPACT_CTOR_DEF, LAMBDA
+MethodLength
+max = 2
+countEmpty = (default)true
+tokens = (default)METHOD_DEF, CTOR_DEF, COMPACT_CTOR_DEF
 
 
 */
 
-//non-compiled with javac: Compilable with Java17
-package com.puppycrawl.tools.checkstyle.checks.sizes.executablestatementcount;
+// Java17
+package com.puppycrawl.tools.checkstyle.checks.sizes.methodlength;
 
-public class InputExecutableStatementCountRecords {
+public class InputMethodLengthRecordsAndCompactCtors {
 
     record MyTestRecord() {
-        static { // violation
+        static {
             System.out.println("test");
             System.out.println("test");
             System.out.println("test");
@@ -21,7 +22,8 @@ public class InputExecutableStatementCountRecords {
 
     //compact ctor
     record MyTestRecord2() {
-        public MyTestRecord2 { // violation
+        // violation below 'Method MyTestRecord2 length is 6 lines (max allowed is 2)'
+        public MyTestRecord2 {
             System.out.println("test");
             System.out.println("test");
             System.out.println("test");
@@ -30,7 +32,7 @@ public class InputExecutableStatementCountRecords {
     }
 
     record MyTestRecord3(String str) {
-        void foo() { // violation
+        void foo() { // violation 'Method foo length is 5 lines (max allowed is 2)'
             System.out.println("test");
             System.out.println("test");
             System.out.println("test");
@@ -38,7 +40,8 @@ public class InputExecutableStatementCountRecords {
     }
 
     record MyTestRecord4(int x, int y) {
-        public MyTestRecord4() { // violation
+        // violation below 'Method MyTestRecord4 length is 7 lines (max allowed is 2)'
+        public MyTestRecord4() {
             this(4, 5);
             System.out.println("test");
             System.out.println("test");
@@ -48,7 +51,7 @@ public class InputExecutableStatementCountRecords {
     }
 
     record MyTestRecord5() {
-        static { // violation
+        static {
             int y = 5;
             int z = 10;
             String newString = String.valueOf(y);
@@ -59,10 +62,10 @@ public class InputExecutableStatementCountRecords {
     }
 
     class LocalRecordHelper {
-        Class<?> m(int x) {
+        Class<?> m(int x) { // violation 'Method m length is 15 lines (max allowed is 2)'
             record R76(int x) {
 
-                public R76 { // violation
+                public R76 { // violation 'Method R76 length is 8 lines (max allowed is 2)'
                     int y = 5;
                     int z = 10;
                     String newString = String.valueOf(y);
@@ -75,20 +78,4 @@ public class InputExecutableStatementCountRecords {
             return R76.class;
         }
     }
-
-    private int id(int i) {
-        return i;
-    }
-
-    private final int value = 2;
-
-    private final int field = id(switch(value) {
-        case 0 -> -1;
-        case 2 -> {
-            int temp = 0;
-            temp += 3;
-            yield temp;
-        }
-        default -> throw new IllegalStateException();
-    });
 }
