+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/sizes/outertypenumber/InputOuterTypeNumberRecords.java
@@ -5,7 +5,7 @@ max = (default)1
 
 */
 
-//non-compiled with javac: Compilable with Java17
+// Java17
 package com.puppycrawl.tools.checkstyle.checks.sizes.outertypenumber; // violation
 
 class InputOuterTypeNumberRecords { }
