+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/sizes/recordcomponentnumber/InputRecordComponentNumberMax1.java
@@ -1,12 +1,12 @@
 /*
 RecordComponentNumber
-max = 20
+max = 1
 accessModifiers = (default)public, protected, package, private
 
 
 */
 
-//non-compiled with javac: Compilable with Java17
+// Java17
 package com.puppycrawl.tools.checkstyle.checks.sizes.recordcomponentnumber;
 
 import java.awt.Point;
@@ -17,7 +17,7 @@ import java.util.List;
 
 import org.w3c.dom.Node;
 
-public class InputRecordComponentNumberMax20 {
+public class InputRecordComponentNumberMax1 {
 
     public record TestRecord1(int x){
         public TestRecord1{
@@ -25,15 +25,15 @@ public class InputRecordComponentNumberMax20 {
         }
     }
 
-    public record TestRecord2(int x, int y){
+    public record TestRecord2(int x, int y){ // violation
 
     }
 
-    public record TestRecord3(String str, int x, int y){
+    public record TestRecord3(String str, int x, int y){ // violation
 
     }
 
-    public record TestRecord4(Node node,
+    public record TestRecord4(Node node, // violation
                               Point x,
                               Point y,
                               int translation,
@@ -49,43 +49,37 @@ public class InputRecordComponentNumberMax20 {
         }
     }
 
-    public record TestRecord5(int x, int y, int z,
+    public record TestRecord5(int x, int y, int z, // violation
                               int a, int b, int c, int d){
 
     }
 
-    public record TestRecord6(int x, int y, int z,
-                              int a, int b, int c,
-                              int d, int e, int f,
-                              int g, int h, int i,
-                              int j, int k){
+    public record TestRecord6(int x, int y, int z, // violation
+                              int a, int b, int c, int d, int e,
+                              int f, int g, int h, int i, int j,
+                              int k){
 
     }
     public record TestRecord7(int y){
 
-        record InnerRecordOk(int x, int y, int z){
+        record InnerRecordOk(int x, int y, int z){ // violation
 
         }
 
-        private record InnerRecordBad(int x, int y, int z,
-                                      int a, int b, int c,
-                                      int d, int e, int f,
-                                      int g, int h, int i,
-                                      int j, int k){
+        private record InnerRecordBad(int x, int y, int z, // violation
+                                      int a, int b, int c, int d, int e,
+                                      int f, int g, int h, int i, int j, int k){
 
-            private record InnerRecordCeptionBad(int x, int y, int z,
-                                                 int a, int b, int c,
-                                                 int d, int e, int f,
-                                                 int g, int h, int i,
-                                                 int j, int k) {
+            private record InnerRecordCeptionBad(int x, int y, int z, // violation
+                                      int a, int b, int c, int d, int e,
+                                      int f, int g, int h, int i, int j, int k){
 
-                public record InnerPublicBad(int[] arr,
+                public record InnerPublicBad(int[] arr, // violation
                                              LinkedHashMap<String, Node> linkedHashMap,
                                              int x,
                                              ArrayDeque<Node> arrayDeque,
                                              List<String> myList,
                                              List<String> myOtherList) {
-
 
                 }
             }
@@ -93,11 +87,9 @@ public class InputRecordComponentNumberMax20 {
         }
     }
 
-    public record TestRecord8(int x, int y, int z, String... myVarargs){
+    public record TestRecord8(int x, int y, int z, String... myVarargs){} // violation
 
-    }
-
-    public record TestRecord9(int x, int y, int z,
+    public record TestRecord9(int x, int y, int z, // violation
                               int a, int b, int c,
                               int d, int e, int f,
                               int g, int h, int i,
@@ -107,8 +99,11 @@ public class InputRecordComponentNumberMax20 {
 
     public record TestRecord10(String... myVarargs){}
 
-    public record TestRecord11(int[] arr, LinkedHashMap<String, Node> linkedHashMap, int x){}
-    public record TestRecord12(int[] arr,
+    public record TestRecord11(int[] arr, // violation
+                               LinkedHashMap<String, Node> linkedHashMap,
+                               int x){}
+
+    public record TestRecord12(int[] arr, // violation
                                LinkedHashMap<String, Node> linkedHashMap,
                                int x,
                                ArrayDeque<Node> arrayDeque,
@@ -119,6 +114,6 @@ public class InputRecordComponentNumberMax20 {
 
     private static record MyPrivateRecord1() {}
 
-    private static record MyPrivateRecord2(int x, int y) {}
+    private static record MyPrivateRecord2(int x, int y) {} // violation
 
 }
