+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/sizes/recordcomponentnumber/InputRecordComponentNumberTopLevel1.java
@@ -6,7 +6,7 @@ accessModifiers = (default)public, protected, package, private
 
 */
 
-//non-compiled with javac: Compilable with Java17
+// Java17
 package com.puppycrawl.tools.checkstyle.checks.sizes.recordcomponentnumber;
 
 public record InputRecordComponentNumberTopLevel1(int x, int y, int z, // violation
